small fixes

+++ src/main/kotlin/fr/lidonis/adventofcode/y2019/day21/SpringDroid.kt
@@ -14,7 +14,7 @@ class SpringDroid(program: String) {
         computer.run()
         val result = computer.outputs.pollLast()
         if (printOutput) {
-            computer.outputs.forEach { print(it.toChar()) }
+            computer.outputs.forEach { print(it.toInt().toChar()) }
         }
         return result
     }
